+++ src/App.tsx
@@ -5,24 +5,9 @@ import { Projects } from "./scenes/projects/projects";
 import { About } from "./scenes/about/about";
 import { Contact } from "./scenes/contact/contact";
 import { Menu } from "./scenes/menu/menu";
-import { NavProvider, useNavContext } from "./data/navProvider";
+import { NavProvider } from "./data/navProvider";
 
 function App() {
-  const { updateScrollPosition } = useNavContext();
-  const handleScroll = () => {
-    let position = window.scrollY;
-    console.log(position);
-    updateScrollPosition(position);
-  };
-
-  useEffect(() => {
-    window.addEventListener("scroll", handleScroll, { passive: true });
-
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, []);
-
   return (
     <NavProvider>
       <div className="App">
