+++ src/scenes/about/about.module.scss
@@ -2,5 +2,5 @@
 
 .container {
     background-color: rgb(184, 66, 11);
-    min-height: 50vh;
+    min-height: 80vh;
 }
