+++ src/scenes/contact/contact.module.scss
@@ -1,4 +1,4 @@
 .container {
     background-color: rgb(184, 11, 132);
-    min-height: 50vh;
+    min-height: 100vh;
 }
